feat: init composer; configure PSR-4 autoload; move classes to src directory; add namespace declaration for User, Post and Comment classes
